Code's finally compiling
Turns out that the accessor for the protobuf fields is what you get if
you take the lowercase of the actual name. so ‘blockId’ - notice the
capital ‘I’ - can be grabbed using .blockid(). Can’t believe I didn’t
notice that till now.

Now to start testing and fixing it up.